pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Nikita Renjov <n.renjov@cft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
mypy = "^1.10.1"
fastapi = "^0.111.1"
python-decouple = "^3.8"
pyjwt = "^2.9.0"
pytest-mock = "^3.14.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.23.8"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "0.18.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
